Move getHighestAssigneeRole from utils to <revision-page> component
