+++ src/ggrc-client/js/components/revision-log/revision-page.js
@@ -7,7 +7,6 @@ import './revision-log-data';
 import {getRolesForType} from '../../plugins/utils/acl-utils';
 import RefreshQueue from '../../models/refresh_queue';
 import template from './revision-page.stache';
-import {getHighestAssigneeRole} from '../../plugins/ggrc_utils';
 import Person from '../../models/business-models/person';
 import Stub from '../../models/stub';
 import Mappings from '../../models/mappers/mappings';
@@ -132,7 +131,7 @@ export default can.Component.extend({
             }
             return {
               updated_at: rev.updated_at,
-              role: getHighestAssigneeRole(
+              role: this.getHighestAssigneeRole(
                 instance,
                 rev.content.attrs.AssigneeType.split(',')),
             };
@@ -176,7 +175,7 @@ export default can.Component.extend({
 
       _.forEach(unmodifiedAssignees, (pid) => {
         let existingRoles = assigneeRoles[pid];
-        let role = getHighestAssigneeRole(
+        let role = this.getHighestAssigneeRole(
           instance, existingRoles);
         perPersonRoleHistory[pid] = [{
           updated_at: instance.created_at,
@@ -641,6 +640,33 @@ export default can.Component.extend({
       );
 
       return peopleList.length ? peopleList : [EMPTY_DIFF_VALUE];
+    },
+    /**
+     * A function that returns the highest role in an array of strings of roles
+     * or a comma-separated string of roles.
+     *
+     * @param {Cacheable} obj - Assignable object with defined
+     *   assignable_list class property holding assignable roles ordered in
+     *   increasing importance.
+     * Return highest assignee role from a list of roles
+     * @param {Array|String} roles - An Array of strings or a String with comma
+     *   separated values of roles.
+     * @return {string} - Highest role from an array of strings or 'none' if
+     *   none found.
+     */
+    getHighestAssigneeRole(obj, roles) {
+      let roleOrder = _.map(
+        _.map(obj.class.assignable_list, 'type'),
+        _.capitalize);
+
+      if (_.isString(roles)) {
+        roles = roles.split(',');
+      }
+
+      roles = _.map(roles, _.capitalize);
+
+      roles.unshift('none');
+      return _.maxBy(roles, Array.prototype.indexOf.bind(roleOrder));
     },
   },
 });
