+++ src/ggrc-client/js/plugins/ggrc_utils.js
@@ -127,34 +127,6 @@ function getPlainText(originalText) {
 }
 
 /**
- * A function that returns the highest role in an array of strings of roles
- * or a comma-separated string of roles.
- *
- * @param {Cacheable} obj - Assignable object with defined
- *   assignable_list class property holding assignable roles ordered in
- *   increasing importance.
- * Return highest assignee role from a list of roles
- * @param {Array|String} roles - An Array of strings or a String with comma
- *   separated values of roles.
- * @return {string} - Highest role from an array of strings or 'none' if
- *   none found.
- */
-function getHighestAssigneeRole(obj, roles) {
-  let roleOrder = _.map(
-    _.map(obj.class.assignable_list, 'type'),
-    _.capitalize);
-
-  if (_.isString(roles)) {
-    roles = roles.split(',');
-  }
-
-  roles = _.map(roles, _.capitalize);
-
-  roles.unshift('none');
-  return _.maxBy(roles, Array.prototype.indexOf.bind(roleOrder));
-}
-
-/**
  * Build string of assignees types separated by commas.
  * @param {Object} instance - Object instance
  * @return {String} assignees types separated by commas
@@ -204,7 +176,6 @@ export {
   loadScript,
   hasPending,
   getPlainText,
-  getHighestAssigneeRole,
   getAssigneeType,
   getTruncatedList,
 };
